Added unit tests

+++ NEATCore/Inc/Parameters.hpp
@@ -1,10 +1,12 @@
 #pragma once
 
+#include <cstddef>
+
 namespace Parameters
 {
    const size_t population_size = 20;
-   const short disjoint_coeff = 2.0;
-   const short excess_coeff = 2.0;
-   const short weights_coeff = 0.4;
-   const short distance_threshold = 1.0;
+   const float disjoint_coeff = 2.0;
+   const float excess_coeff = 2.0;
+   const float weights_coeff = 0.4;
+   const float distance_threshold = 1.0;
 }
